Legal pages reworks
