--- frontend/components/PrivacyPolicy/index.module.scss
@@ -1,77 +1,0 @@
-@import '../../styles/variables';
-@import '../../styles/common';
-
-.privacyPolicyWrapper {
-  padding: $gapS;
-  margin: 0 auto;
-  max-width: 800px;
-  font-size: 16px;
-  line-height: 1.5;
-
-  @media screen and (min-width: $desktop) {
-    padding: $gapM;
-  }
-
-  h1 {
-    font-size: 24px;
-    font-weight: 600;
-    margin-bottom: $gapS;
-  }
-
-  h2 {
-    font-size: 20px;
-    font-weight: 600;
-    margin-bottom: $gapS;
-  }
-
-  h3 {
-    font-size: 18px;
-    font-weight: 400;
-    color: #4155f5;
-    margin-bottom: $gapS;
-  }
-
-  h4 {
-    font-size: 16px;
-    font-weight: 600;
-    margin-bottom: $gapS;
-  }
-
-  h5 {
-    font-size: 14px;
-    font-weight: 600;
-    margin-bottom: $gapS;
-  }
-
-  h6 {
-    font-size: 12px;
-    font-weight: 600;
-    margin-bottom: $gapS;
-  }
-
-  p {
-    margin-bottom: $gapS;
-  }
-
-  a {
-    text-decoration: underline;
-  }
-
-  ul {
-    padding-left: $gapS;
-    list-style-type: disc;
-    margin-bottom: $gapS;
-  }
-
-  li {
-    margin-bottom: $gapS;
-  }
-
-  ol {
-    margin-bottom: $gapS;
-  }
-
-  blockquote {
-    margin-bottom: $gapS;
-  }
-}
